+++ R/utils.R
@@ -1,7 +1,3 @@
-is_Date <- function(x) {
-  inherits(x, 'Date')
-}
-
 is_list <- function(x) {
   class(x) == 'list'
 }
@@ -16,47 +12,120 @@ cdr <- function(cons) {
   cons[-1]
 }
 
-default <- function(x) {
+names2 <- function(x) {
+  if (is.null(names(x))) rep.int("", length(x)) else names(x)
+}
+
+#
+# the default attribute is used by `variables()` and `pair_off()` to know when
+# to assign a variable its default value
+#
+get_default <- function(x) {
   attr(x, "default", exact = TRUE)
 }
 
 has_default <- function(x) {
-  vapply(x, function(i) !is.null(attr(i, "default")), logical(1))
+  vapply(x, function(i) !is.null(get_default(i)), logical(1))
 }
 
+#
+# append any default values onto the end of a list of values, used in
+# `pair_off()` to extend the current set of values thereby avoiding an
+# incorrect number of values error
+#
 add_defaults <- function(names, values, env) {
   where <- which(has_default(names))
-  defaults <- lapply(names[where], default)[where > length(values)]
+  defaults <- lapply(names[where], get_default)[where > length(values)]
   evaled <- lapply(defaults, eval, envir = env)
 
   append(values, evaled)
 }
 
-names2 <- function(x) {
-  if (is.null(names(x))) rep.int("", length(x)) else names(x)
+#
+# traverse nested extract op calls to find the extractee, e.g. `x[[1]][[1]]`
+#
+traverse_to_extractee <- function(call) {
+  if (is.language(call) && is.symbol(call)) {
+    return(call)
+  }
+  traverse_to_extractee(call[[2]])
 }
 
+#
+# used by multi_assign to confirm all extractees exist
+#
+check_extract_calls <- function(lhs, envir) {
+  if (is.character(lhs)) {
+    return()
+  }
+
+  if (is.language(lhs)) {
+    extractee <- traverse_to_extractee(lhs)
+    if (!exists(as.character(extractee), envir = envir, inherits = FALSE)) {
+      stop_invalid_lhs(object_does_not_exist(extractee))
+    } else {
+      return()
+    }
+  }
+
+  unlist(lapply(lhs, check_extract_calls, envir = envir))
+}
+
+is_extract_op <- function(x) {
+  if (length(x) < 1) {
+    return(FALSE)
+  }
+
+  (as.character(x) %in% c("[", "[[", "$"))
+}
+
+is_valid_call <- function(x) {
+  if (length(x) < 1) {
+    return(FALSE)
+  }
+
+  (x == "c" || x == "=" || is_extract_op(x))
+}
+
+#
+# used by multi_assign to assign list elements in the calling environment
+#
+assign_extract <- function(call, value, envir = parent.frame()) {
+  replacee <- call("<-", call, value)
+  eval(replacee, envir = envir)
+  invisible(value)
+}
+
+#
+# parses a substituted expression to create a tree-like list structure,
+# perserves calls to extract ops instead of converting them to lists
+#
 tree <- function(x) {
-  if (length(x) == 1 && is.language(x) && !is.symbol(x)) {
+  if (length(x) == 1) {
     return(x)
   }
 
-  x <- as.list(x)
-
-  if (length(x) == 1 && length(x[[1]]) <= 1) {
-    if (names2(x) != "") {
-      return(list(as.symbol("="), as.symbol(names(x)), x[[1]]))
-    }
-
-    return(x[[1]])
+  if (is_extract_op(x[[1]])) {
+    return(x)
   }
 
-  append(
-    tree(x[[1]]),
-    lapply(seq_along(x[-1]), function(i) tree(x[-1][i]))
+  lapply(
+    seq_along(as.list(x)),
+    function(i) {
+      if (names2(x[i]) != "") {
+        return(list(as.symbol("="), names2(x[i]), x[[i]]))
+      } else {
+        tree(x[[i]])
+      }
+    }
   )
 }
 
+#
+# given a tree-like list structure returns a character vector of the function
+# calls, used by multi_assign to determine if performing standard assignment or
+# multiple assignment
+#
 calls <- function(x) {
   if (!is_list(x)) {
     return(NULL)
@@ -64,17 +133,25 @@ calls <- function(x) {
 
   this <- car(x)
 
-  if (this != "c" && this != "=") {
+  if (!is_valid_call(this)) {
     stop_invalid_lhs(unexpected_call(this))
   }
 
   c(as.character(this), unlist(lapply(cdr(x), calls)))
 }
 
+#
+# given a tree-like list structure, returns a nested list of the variables
+# in the tree, will also associated default values with variables
+#
 variables <- function(x) {
   if (!is_list(x)) {
     if (x == "") {
       stop_invalid_lhs(empty_variable(x))
+    }
+
+    if (is.language(x) && length(x) > 1 && is_extract_op(x[[1]])) {
+      return(x)
     }
 
     if (!is.symbol(x)) {
@@ -108,6 +185,10 @@ incorrect_number_of_values <- function() {
   "incorrect number of values"
 }
 
+object_does_not_exist <- function(obj) {
+  paste0("object `", obj, "` does not exist in calling environment")
+}
+
 empty_variable <- function(obj) {
   paste("found empty variable, check for extraneous commas")
 }
@@ -137,4 +218,8 @@ stop_invalid_lhs <- function(message, call = sys.call(-1), ...) {
 stop_invalid_rhs <- function(message, call = sys.call(-1), ...) {
   cond <- condition(c("invalid_rhs", "error"), message, call, ...)
   stop(cond)
+}
+
+is_invalid_side_error <- function(e) {
+  inherits(e, c("invalid_lhs", "invalid_rhs"))
 }