+++ tests/testthat/test-utils.R
@@ -1,9 +1,5 @@
 context(" * testing utils")
 
-test_that("is_Date returns TRUE for date objects", {
-  expect_true(is_Date(Sys.Date()))
-})
-
 test_that("is_list returns TRUE for list object", {
   expect_true(is_list(list(1, 2, 3)))
 })
@@ -25,4 +21,12 @@ test_that("car throws error for non-list or 0 length list", {
 test_that("car returns first element of list", {
   expect_equal(car(list(1, 2)), 1)
   expect_equal(car(list(list(1, 2), 3)), list(1, 2))
+})
+
+test_that("traverse_to_extractee gets flat and nested extractees", {
+  s1 <- substitute(x[[1]])
+  expect_equal(as.character(traverse_to_extractee(s1)), "x")
+
+  s2 <- substitute(y[[1]][[3030]])
+  expect_equal(as.character(traverse_to_extractee(s2)), "y")
 })